【CE】[AtCoder Beginner Contest 179]F - Simplified Reversi

+++ abc179/abc179_f.cpp
@@ -110,7 +110,7 @@ ll solve(ll n, vector<Query> &queries) {
     }
 
 
-    int mi = n - 1;
+    ll mi = n - 1;
 
     vector<ll> ups(n + 10, -1);
     ups[0] = n - 1;
